fix: explicitly assign problematic portal interfaces (that fallback to gnome) to regolith portal
Portals may not follow the expected ordering as per [this](https://github.com/flatpak/xdg-desktop-portal/issues/1111) issue.

This causes some of the interfaces that are made available by other backend (like gtk or wlr) to get assigned to the no-op
implementations of the xdp-regolith backend.

This PR explicitly assigns only the problematic interfaces, i.e. the ones that get assigned to xdp-gnome, to the regolith backend.